+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+<a name="1.2.0"></a>
+# [1.2.0](https://github.com/flextype-plugins/admin) (2020-08-19)
+
+### Features
+
+* **core** update code base for new Flextype 0.9.10
+
+<a name="1.1.1"></a>
+# [1.1.1](https://github.com/flextype-plugins/admin) (2020-08-xx)
+
+### Features
+* **lang** add new translates
+* **plugins** show additional plugins information
+
+### Bug Fixes
+
+* **plugins-controller** fix plugins repository link
+* **entries-controller** fix method move()
+
 <a name="1.1.0"></a>
 # [1.1.0](https://github.com/flextype-plugins/admin) (2020-08-05)